requirements.txt
fastapi
uvicorn
pydantic
python-multipart
langchain-community
langchain-text-splitters
langchain-huggingface
chromadb
pypdf
sentence-transformers
python-dotenv
serpapi
requests
